+++ ragavi/ragavi.py
@@ -3,6 +3,7 @@ import glob
 import numpy as np
 import re
 import logging
+import warnings
 
 from datetime import datetime
 import matplotlib.cm as cmx
@@ -12,19 +13,69 @@ from argparse import ArgumentParser
 import matplotlib.colors as colors
 
 from bokeh.plotting import figure
-from bokeh.models.widgets import Div
+from bokeh.models.widgets import Div, PreText
 from bokeh.layouts import row, column, gridplot, widgetbox
 from bokeh.io import (output_file, show, output_notebook, export_svgs,
                       export_png, save)
 from bokeh.models import (Range1d, HoverTool, ColumnDataSource, LinearAxis,
                           BasicTicker, Legend, Toggle, CustomJS, Title,
-                          CheckboxGroup, Select, Text)
+                          CheckboxGroup, Select, Text, Slider)
 
-logfile_name = 'ragavi.log'
-logging.basicConfig(filename=logfile_name, filemode='a',
-                    format='%(asctime)s %(message)s',
-                    datefmt='%d.%m.%Y @ %I:%M:%S',
-                    level=logging.DEBUG)
+
+PLOT_WIDTH = 700
+PLOT_HEIGHT = 600
+GAIN_TYPES = ['B', 'F', 'G', 'K']
+GHZ = 1e9
+
+
+def config_logger():
+    """This function is used to configure the logger for ragavi and catch
+        all warnings output by sys.stdout.
+    """
+    logfile_name = 'ragavi.log'
+    # capture only a single instance of a matching repeated warning
+    warnings.filterwarnings('default')
+
+    # setting the format for the logging messages
+    start = " (O_o) ".center(80, "=")
+    form = '{}\n%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+    form = form.format(start)
+    formatter = logging.Formatter(form, datefmt='%d.%m.%Y@%H:%M:%S')
+
+    # setup for ragavi logger
+    logger = logging.getLogger(__name__)
+    logger.setLevel(logging.DEBUG)
+
+    # capture all stdout warnings
+    logging.captureWarnings(True)
+    warnings_logger = logging.getLogger('py.warnings')
+    warnings_logger.setLevel(logging.DEBUG)
+
+    # setup for logfile handing ragavi
+    fh = logging.FileHandler(logfile_name)
+    fh.setLevel(logging.INFO)
+    fh.setFormatter(formatter)
+
+    logger.addHandler(fh)
+    warnings_logger.addHandler(logger)
+    return logger
+
+
+def _handle_uncaught_exceptions(extype, exval, extraceback):
+    """Function to Capture all uncaught exceptions into the log file
+
+       Inputs to this function are acquired from sys.excepthook. This
+       is because this function overrides sys.excepthook 
+
+       https://docs.python.org/3/library/sys.html#sys.excepthook
+
+    """
+    message = "Oops ... !"
+    logger.error(message, exc_info=(extype, exval, extraceback))
+
+
+logger = config_logger()
+sys.excepthook = _handle_uncaught_exceptions
 
 
 def save_svg_image(img_name, figa, figb, glax1, glax2):
@@ -179,12 +230,12 @@ def make_plots(source, ax1, ax2, color='purple', y1_err=None, y2_err=None):
         Tuple of glyphs
 
     """
-    p1 = ax1.circle('x', 'y1', size=8, alpha=1, color=color, source=source,
+    p1 = ax1.circle('x', 'y1', size=4, alpha=1, color=color, source=source,
                     nonselection_color='#7D7D7D', nonselection_fill_alpha=0.3)
     p1_err = errorbar(fig=ax1, x=source.data['x'], y=source.data['y1'],
                       color=color, yerr=y1_err)
 
-    p2 = ax2.circle('x', 'y2', size=8, alpha=1, color=color, source=source,
+    p2 = ax2.circle('x', 'y2', size=4, alpha=1, color=color, source=source,
                     nonselection_color='#7D7D7D', nonselection_fill_alpha=0.3)
     p2_err = errorbar(fig=ax2, x=source.data['x'], y=source.data['y2'],
                       color=color, yerr=y2_err)
@@ -403,6 +454,25 @@ def legend_toggle_callback():
     return code
 
 
+def size_slider_callback():
+    """JS callback to sel
+    """
+
+    code = """
+            
+            var pos, i, numplots;
+
+            numplots = p1.length;
+            pos = slide.value;
+
+            for (i=0; i<numplots; i++){
+                p1[i].glyph.size = pos;
+                p2[i].glyph.size = pos;
+            }
+           """
+    return code
+
+
 def create_legend_batches(num_leg_objs, li_ax1, li_ax2, lierr_ax1, lierr_ax2, batch_size=16):
     """Automates creation of antenna batches of 16 each unless otherwise
 
@@ -499,10 +569,12 @@ def create_legend_objs(num_leg_objs, bax1, baerr_ax1, bax2, baerr_ax2):
     for i in range(num_leg_objs):
         lo_ax1['leg_%s' % str(i)] = Legend(items=bax1[i],
                                            location='top_right',
-                                           click_policy='hide')
+                                           click_policy='hide',
+                                           visible=False)
         lo_ax2['leg_%s' % str(i)] = Legend(items=bax2[i],
                                            location='top_right',
-                                           click_policy='hide')
+                                           click_policy='hide',
+                                           visible=False)
         loerr_ax1['leg_%s' % str(i)] = Legend(
             items=baerr_ax1[i],
             location='top_right',
@@ -615,7 +687,7 @@ def name_2id(val, dic):
         return -1
 
 
-def data_prep_G(masked_data, masked_data_err, doplot, corr):
+def data_prep_G(masked_data, masked_data_err, doplot):
     """Preparing the data for plotting gain cal-table
 
     Inputs
@@ -626,8 +698,7 @@ def data_prep_G(masked_data, masked_data_err, doplot, corr):
         Flagged data from the PARAMERR column to be plotted
     doplot: str
         Either 'ap' or 'ri'
-    corr: int
-        Correlation to plot (0,1 e.t.c)
+
 
     Outputs
     -------
@@ -637,22 +708,22 @@ def data_prep_G(masked_data, masked_data_err, doplot, corr):
     """
 
     if doplot == 'ap':
-        y1 = np.abs(masked_data)[:, 0, corr]
-        y1_err = np.abs(masked_data_err)[:, 0, corr]
-        y2 = np.angle(masked_data, deg=True)[:, 0, corr]
+        y1 = np.ma.abs(masked_data)
+        y1_err = np.ma.abs(masked_data_err)
+        y2 = np.ma.angle(masked_data, deg=True)
         # Remove phase limit from -pi to pi
-        y2 = np.unwrap(y2)
+        #y2 = np.unwrap(y2)
         y2_err = None
     else:
-        y1 = np.real(masked_data)[:, 0, corr]
-        y1_err = np.abs(masked_data_err)[:, 0, corr]
-        y2 = np.imag(masked_data)[:, 0, corr]
+        y1 = np.real(masked_data)
+        y1_err = np.ma.abs(masked_data_err)
+        y2 = np.imag(masked_data)
         y2_err = None
 
     return y1, y1_err, y2, y2_err
 
 
-def data_prep_B(masked_data, masked_data_err, doplot, corr):
+def data_prep_B(masked_data, masked_data_err, doplot):
     """Preparing the data for plotting bandpass cal-table
 
     Inputs
@@ -663,8 +734,6 @@ def data_prep_B(masked_data, masked_data_err, doplot, corr):
         Flagged data from the PARAMERR column to be plotted
     doplot: str
         Either 'ap' or 'ri'
-    corr: int
-        Correlation to plot (0,1 e.t.c)
 
     Outputs
     -------
@@ -672,15 +741,12 @@ def data_prep_B(masked_data, masked_data_err, doplot, corr):
         Tuple with arrays of the different data
 
     """
-    masked_data = masked_data[:, :, corr]
-    masked_data_err = masked_data_err[:, :, corr]
-
     if doplot == 'ap':
-        y1 = np.abs(masked_data)
-        y1_err = np.abs(masked_data_err)
-        y2 = np.array(np.angle(masked_data, deg=True))
-        y2 = np.unwrap(y2)
-        y2_err = np.array(np.angle(masked_data_err, deg=True))
+        y1 = np.ma.abs(masked_data)
+        y1_err = np.ma.abs(masked_data_err)
+        y2 = np.ma.angle(masked_data, deg=True)
+        #y2 = np.unwrap(y2)
+        y2_err = np.ma.angle(masked_data_err, deg=True)
         y2_err = np.unwrap(y2_err)
     else:
         y1 = np.real(masked_data)
@@ -691,7 +757,7 @@ def data_prep_B(masked_data, masked_data_err, doplot, corr):
     return y1, y1_err, y2, y2_err
 
 
-def data_prep_K(masked_data, masked_data_err, corr):
+def data_prep_K(masked_data, masked_data_err, doplot):
     """Preparing the data for plotting delay cal-table. Doplot must be 'ap'.
 
     Inputs
@@ -700,8 +766,6 @@ def data_prep_K(masked_data, masked_data_err, corr):
         Flagged data from CPARAM column to be plotted.
     masked_data_err: numpy.ndarray
         Flagged data from the PARAMERR column to be plotted
-    corr: int
-        Correlation to plot (0,1 e.t.c)
 
     Outputs
     -------
@@ -709,8 +773,7 @@ def data_prep_K(masked_data, masked_data_err, corr):
         Tuple with arrays of the different data
 
     """
-    y1 = masked_data[:, 0, corr]
-    y1 = np.array(y1)
+    y1 = masked_data
     y1_err = masked_data_err
 
     # quick fix to hide plot without generating errors
@@ -720,7 +783,7 @@ def data_prep_K(masked_data, masked_data_err, corr):
     return y1, y1_err, y2, y2_err
 
 
-def data_prep_F(masked_data, masked_data_err, doplot, corr):
+def data_prep_F(masked_data, masked_data_err, doplot):
     """Preparing the data for plotting flux cal table
 
     Inputs
@@ -731,8 +794,6 @@ def data_prep_F(masked_data, masked_data_err, doplot, corr):
         Flagged data from the PARAMERR column to be plotted
     doplot: str
         Either 'ap' or 'ri'
-    corr: int
-        Correlation to plot (0,1 e.t.c)
 
     Outputs
     -------
@@ -742,19 +803,383 @@ def data_prep_F(masked_data, masked_data_err, doplot, corr):
     """
 
     if doplot == 'ap':
-        y1 = np.abs(masked_data)[:, 0, corr]
-        y1_err = np.abs(masked_data_err)[:, 0, corr]
-        y2 = np.angle(masked_data, deg=True)[:, 0, corr]
+        y1 = np.ma.abs(masked_data)
+        y1_err = np.ma.abs(masked_data_err)
+        y2 = np.ma.angle(masked_data, deg=True)
         # Remove phase limit from -pi to pi
-        y2 = np.unwrap(y2)
+        #y2 = np.unwrap(y2)
         y2_err = None
     else:
-        y1 = np.real(masked_data)[:, 0, corr]
-        y1_err = np.abs(masked_data_err)[:, 0, corr]
-        y2 = np.imag(masked_data)[:, 0, corr]
+        y1 = np.real(masked_data)
+        y1_err = np.ma.abs(masked_data_err)
+        y2 = np.imag(masked_data)
         y2_err = None
+    return y1, y1_err, y2, y2_err
+
+
+def get_yaxis_data(table_obj, gtype, ptype):
+    """ Function to extract the required column for the y-axis data.
+    This column is determined by ptype which can be amplitude vs phase 'ap'
+    or real vs imaginary 'ri'.
+
+    Inputs
+    -----
+    table_obj: python casacore table object
+               Table in which to get the data
+
+    gtype: str
+           Gain table type B, F, G or K.
+
+    ptype: str
+           Plot type ap / ri
+
+    Outputs
+    -------
+    Returns np.ndarray data as well as the y-axis labels (str) for both plots.
+    """
+    if ptype == 'ap':
+        y1_label = 'Amplitude'
+        y2_label = 'Phase[deg]'
+    else:
+        y1_label = 'Real'
+        y2_label = 'Imaginary'
+
+    if gtype == 'K':
+        data_column = 'FPARAM'
+    else:
+        data_column = 'CPARAM'
+
+    ydata = table_obj.getcol(data_column)
+    return ydata, y1_label, y2_label
+
+
+def prep_yaxis_data(table_obj, ydata, gtype, ptype='ap', corr=0, flag=True):
+    """Function to process data for the y-axis. Part of the processing includes:
+    - Selecting correlation for the data and error
+    - Flagging
+    - Complex correlation parameter conversion to amplitude, phase, real and 
+      imaginary for processing
+    Data selection and flagging are done by this function itself, however ap and ri conversion are done by specified functions.
+
+    Inputs
+    ------
+    table_obj: pyrap table object
+               table object for an already open table
+    ydata: ndarray
+           Relevant y-axis data to be processed
+    gtype: str
+           Gain table type  B, F, G or K.
+    ptype: str
+           Plot type 'ap' / 'ri'
+    corr: int
+          Correlation number to select
+    flag: bool
+          Option on whether to flag the data or not
+
+    Outputs
+    -------
+    y1: masked ndarray
+        Amplitude / real part of the complex input data.
+    y1_err: masked ndarray
+        Error data for y1.
+    y2: masked ndarray
+        Phase angle / Imaginary part of input data.
+    y2_err: masked ndarray
+        Error data for y2.
+
+    """
+    # select data correlation for both the data and the errors
+    ydata = ydata[:, :, corr]
+    ydata_errors = get_errors(table_obj)[:, :, corr]
+
+    if flag:
+        flags = get_flags(table_obj, corr)
+        ydata = np.ma.masked_array(data=ydata, mask=flags)
+        ydata_errors = np.ma.masked_array(data=ydata_errors, mask=flags)
+
+    if gtype == 'B':
+        y1, y1_err, y2, y2_err = data_prep_B(ydata, ydata_errors, ptype)
+    elif gtype == 'F':
+        ydata = ydata[:, 0]
+        ydata_errors = ydata_errors[:, 0]
+        y1, y1_err, y2, y2_err = data_prep_F(ydata, ydata_errors, ptype)
+    elif gtype == 'G':
+        ydata = ydata[:, 0]
+        ydata_errors = ydata_errors[:, 0]
+        y1, y1_err, y2, y2_err = data_prep_G(ydata, ydata_errors, ptype)
+    elif gtype == 'K':
+        ydata = ydata[:, 0]
+        ydata_errors = ydata_errors[:, 0]
+        y1, y1_err, y2, y2_err = data_prep_K(ydata, ydata_errors, ptype)
 
     return y1, y1_err, y2, y2_err
+
+
+def get_frequencies(table_obj):
+    """Function to get channel frequencies from the SPECTRAL_WINDOW subtable.
+    Inputs
+    ------
+    table_obj: pyrap table object
+
+    Outputs
+    -------
+    freqs: 1D-array
+           Channel centre frequencies.
+    """
+    spw_subtab = table(table_obj.getkeyword('SPECTRAL_WINDOW'), ack=False)
+    freqs = spw_subtab.getcell('CHAN_FREQ', 0)
+    spw_subtab.close()
+    return freqs
+
+
+def get_antennas(table_obj):
+    """Function to get antennae names from the ANTENNA subtable.
+    Inputs
+    ------
+    table_obj: pyrap table object
+
+    Outputs
+    -------
+    ant_names: 1D-array
+               Names for all the antennas available.
+
+    """
+    ant_subtab = table(table_obj.getkeyword('ANTENNA'), ack=False)
+    ant_names = ant_subtab.getcol('NAME')
+    ant_subtab.close()
+    return ant_names
+
+
+def get_flags(table_obj, corr=None):
+    """Function to get Flag values from the FLAG column
+    Allows the selection of flags for a single correlation. If none is specified the entire data is then selected.
+    Inputs
+    ------
+    table_obj: pyrap table object
+    corr: int
+          Correlation number to select.
+
+    Outputs
+    -------
+    flags: ndarray
+           Array containing selected flag values.
+
+    """
+    flags = table_obj.getcol('FLAG')
+    if corr is None:
+        return flags
+    else:
+        flags = flags[:, :, corr]
+    return flags
+
+
+def get_errors(table_obj):
+    """Function to get error data from PARAMERR column.
+    Inputs
+    ------
+    table_obj: pyrap table object.
+
+    Outputs
+    errors: ndarray
+            Error data. 
+    """
+    errors = table_obj.getcol('PARAMERR')
+    return errors
+
+
+def get_fields(table_obj):
+    """Function to get field names from the FIELD subtable.
+    Inputs
+    ------
+    table_obj: pyrap table object
+
+    Outputs
+    -------
+    field_names: 1-D array
+                 String names for the available data in the table
+    """
+    field_subtab = table(table_obj.getkeyword('FIELD'), ack=False)
+    field_names = field_subtab.getcol('NAME')
+    field_subtab.close()
+    return field_names
+
+
+def get_tooltip_data(table_obj, gtype):
+    """Function to get the data to be displayed on the mouse tooltip on the plots.
+    Inputs
+    ------
+    table_obj: pyrap table object
+    gtype: str
+           Type of gain table being plotted
+
+    Outputs
+    -------
+    spw_id: ndarray
+            Spectral window ids
+    scan_no: ndarray
+             scan ids
+    ttip_antnames: array
+                   Antenna names to show up on the tooltips
+
+
+    """
+    spw_id = table_obj.getcol('SPECTRAL_WINDOW_ID')
+    scan_no = table_obj.getcol('SCAN_NUMBER')
+    ant_id = table_obj.getcol('ANTENNA1')
+    antnames = get_antennas(table_obj)
+    # get available antenna names from antenna id
+    ttip_antnames = np.array([antnames[x] for x in ant_id])
+
+    freqs = get_frequencies(table_obj)
+    nchan = len(freqs)
+
+    if gtype == 'B':
+        spw_id = spw_id.reshape(spw_id.size, 1)
+        spw_id = spw_id.repeat(nchan, axis=1)
+        scan_no = scan_no.reshape(scan_no.size, 1)
+        scan_no = scan_no.repeat(nchan, axis=1)
+        ant_id = ant_id.reshape(ant_id.size, 1)
+        ant_id = ant_id.repeat(nchan, axis=1)
+        # ttip_antnames = ttip_antnames.
+
+    return spw_id, scan_no, ttip_antnames
+
+
+def get_xaxis_data(table_obj, gtype):
+    """Function to get x-axis data. It is dependent on the gaintype.
+        This function also returns the relevant x-axis labels for both pairs of plots.
+    Inputs
+    ------
+    table_obj: pyrap table object
+
+    gtype:  str
+            Type of gain table being plotted.
+
+    Outputs
+    -------
+    xdata: ndarray
+           X-axis data depending on the gain table to be plotted.
+    xaxis_label: str
+                 Label to appear on the x-axis of the plots. This is shared amongst both plots.
+    """
+    if gtype == 'B':
+        xdata = get_frequencies(table_obj)
+        xaxis_label = 'Channel'
+    elif gtype == 'F' or gtype == 'G':
+        xdata = table_obj.getcol('TIME')
+        xaxis_label = 'Time[s]'
+    elif gtype == 'K':
+        xdata = table_obj.getcol('ANTENNA1')
+        xaxis_label = 'Antenna'
+
+    return xdata, xaxis_label
+
+
+def prep_xaxis_data(xdata, gtype):
+    """Function to Prepare the x-axis data.
+    Inputs
+    ------
+    xdata: 1-D array
+           Data for the xaxis to be prepared
+    gtype: str
+           gain type of the table to plot
+    ptype: str
+           Type of plot, whether ap or ri
+
+    Outputs
+    -------
+    prepdx: 1-D array
+            Data for the x-axid of the plots
+    """
+    if gtype == 'B':
+        prepdx = np.arange(xdata.size)
+    elif gtype == 'G' or gtype == 'F':
+        prepdx = xdata - xdata[0]
+    elif gtype == 'K':
+        prepdx = xdata
+    return prepdx
+
+
+def stats_display(table_obj, gtype, ptype, corr, field):
+    """Function to display some statistics on the plots. These statistics are derived from a specific correlation and a specified field of the data.
+    Currently, only the medians of these plots are displayed.
+
+    Inputs
+    ------
+    table_obj: pyrap table object
+    gtype: str
+           Type of gain table to be plotted.
+    ptype: str
+            Type of plot ap / ri
+    corr: int
+          Correlation number of the data to be displayed
+    field: int
+            Integer field id of the field being plotted. If a string name was provided, it is converted within the main function by the name2id function.
+
+    Outputs
+    -------
+    pre: Bokeh widget model object
+         Preformatted text containing the medians for both model. The object returned must then be placed within the widget box for display.
+
+    """
+    subtable = table_obj.query(query="FIELD_ID=={}".format(field))
+    ydata, y1label, y2label = get_yaxis_data(subtable, gtype, ptype)
+
+    flags = get_flags(subtable)[:, :, corr]
+    ydata = ydata[:, :, corr]
+    m_ydata = np.ma.masked_array(data=ydata, mask=flags)
+
+    if ptype == 'ap':
+        y1 = np.ma.abs(m_ydata)
+        y2 = np.ma.angle(m_ydata, deg=True)
+        med_y1 = np.ma.median(y1)
+        med_y2 = np.ma.median(y2)
+        text = "Median Amplitude: {}\nMedian Phase: {} deg".format(
+            med_y1, med_y2)
+        if gtype == 'K':
+            text = "Median Amplitude: {}".format(med_y1)
+    else:
+        y1 = np.ma.real(m_ydata)
+        y2 = np.ma.imag(m_ydata)
+
+        med_y1 = np.ma.median(y1)
+        med_y2 = np.ma.median(y2)
+        text = "Median Real: {}\nMedian Imaginary: {}".format(med_y1, med_y2)
+
+    pre = PreText(text=text)
+
+    return pre
+
+
+def autofill_gains_fields(t, g, f):
+    """Normalise length of f and g lists to the length of
+       t list. This function is meant to support  the ability to specify multiple gain tables while only specifying single values for field ids and gain table types. An assumption will be made that for all the specified tables, the same field id and gain table type will be used.
+
+    Inputs
+    ------
+    t: list
+          list of the gain tables.
+    f: str
+            field id to be plotted.
+    g: list 
+           type of gain table [B,G,K,F].
+
+
+    Outputs
+    -------
+    f, g: list
+                    lists of length lengthof(t) containing field ids and gain types.
+    """
+    ltab = len(t)
+    lfields = len(f)
+    lgains = len(g)
+
+    if ltab != lgains and lgains == 1:
+        g = g * ltab
+    if ltab != lfields and lfields == 1:
+        f = f * ltab
+
+    return g, f
 
 
 def get_argparser():
@@ -853,63 +1278,54 @@ def main(**kwargs):
         mytabs = kwargs.get('mytabs', [])
         gain_types = kwargs.get('gain_types', [])
 
-    if len(mytabs) > 0:
-            # getting the name of the gain table specified
-        mytabs = [x.rstrip("/") for x in mytabs]
-    else:
-        logging.info('ragavi exited: No gain table specified.')
+    if len(mytabs) == 0:
+        logger.error('Exiting: No gain table specified.')
         sys.exit(-1)
 
-    # Keep this for notebook rendering
-    if len(gain_types) > 0:
-        GAIN_TYPES = ['B', 'F', 'G', 'K']
-        for gain_type in gain_types:
-            if gain_type.upper() not in GAIN_TYPES:
-                logging.info("Choose appropriate gain_type: ", GAIN_TYPES)
-                sys.exit(-1)
-    else:
-        logging.info('ragavi exited: No gain type specifed.')
+    mytabs = [x.rstrip("/") for x in mytabs]
+
+    if len(gain_types) == 0:
+        logger.error('Exiting: No gain type specified.')
         sys.exit(-1)
+
+    gain_types = [x.upper() for x in gain_types]
 
     if len(field_ids) == 0:
-        loggging.info('ragavi exited: No field id specified.')
+        logger.error('Exiting: No field id specified.')
         sys.exit(-1)
 
+    if doplot not in ['ap', 'ri']:
+        logger.error('Exiting: Plot selection must be ap or ri.')
+        sys.exit(-1)
+
+    # for notebook
+    for gain_type in gain_types:
+        if gain_type not in GAIN_TYPES:
+            logger.error("Exiting: gtype {} invalid".format(gain_type))
+            sys.exit(-1)
+
+    gain_types, field_ids = autofill_gains_fields(mytabs, gain_types,
+                                                  field_ids)
     # array to store final output image
     final_layout = []
+
     for mytab, gain_type, field in zip(mytabs, gain_types, field_ids):
+
         # reinitialise plotant list for each table
         if NB_RENDER:
-            plotants = plotants = kwargs.get('plotants', [-1])
+            plotants = kwargs.get('plotants', [-1])
         else:
             plotants = options.plotants
 
-        # by default is ap: amplitude and phase
-        if doplot not in ['ap', 'ri']:
-            logging.info('ragavi exited: Plot selection must be ap or ri.')
-            sys.exit(-1)
-        # configuring the plot dimensions
-        PLOT_WIDTH = 700
-        PLOT_HEIGHT = 600
+        tt = table(mytab, ack=False)
 
-        # get main table and useful subtables
-        try:
-            tt = table(mytab, ack=False)
-        except RuntimeError as runtime:
-            logging.exception('ragavi exited: Runtime error encountered.')
-            sys.exit(-1)
-        spw_table = table(mytab + '::SPECTRAL_WINDOW', ack=False)
-        field_tab = table(mytab + '::FIELD', ack=False)
-        anttab = table(mytab + '::ANTENNA', ack=False)
-
-        # get columns from selected tables
-        field_names = field_tab.getcol('NAME')
+        field_names = get_fields(tt)
         field_src_ids = dict(enumerate(field_names))
-        antnames = anttab.getcol('NAME')
+        antnames = get_antennas(tt)
         ants = np.unique(tt.getcol('ANTENNA1'))
         fields = np.unique(tt.getcol('FIELD_ID'))
-        flags = tt.getcol('FLAG')
-        frequencies = spw_table.getcol('CHAN_FREQ')[0] / 1e9
+
+        frequencies = get_frequencies(tt) / GHZ
 
         # setting up colors for the antenna plots
         cNorm = colors.Normalize(vmin=0, vmax=len(ants) - 1)
@@ -922,8 +1338,9 @@ def main(**kwargs):
             field = name_2id(field, field_src_ids)
 
         if int(field) not in fields.tolist():
-            logging.info('ragavi exited: Field id {} not found.'.format(field))
-            sys.exit(-1)
+            logger.info(
+                'Skipping table: {} : Field id {} not found.'.format(mytab, field))
+            continue
 
         if plotants[0] != -1:
             # creating a list for the antennas to be plotted
@@ -932,10 +1349,11 @@ def main(**kwargs):
             for ant in plotants:
                 if int(ant) not in ants:
                     plotants.remove(ant)
-                    logging.info(
-                        'Requested antenna ID {} not found.'.format(ant))
+                    logger.info('Antenna ID {} not found.'.format(ant))
+
+            # check if plotants still has items
             if len(plotants) == 0:
-                logging.info('ragavi exited: No valid antennas requested')
+                logger.error('Exiting: No valid antennas requested')
                 sys.exit(-1)
             else:
                 plotants = np.array(plotants, dtype=int)
@@ -943,12 +1361,18 @@ def main(**kwargs):
             plotants = ants
 
         # creating bokeh figures for plots
-        # linking plots ax1 and ax2 via the x_axes because fo similarities in
+        # linking plots ax1 and ax2 via the x_axes because of similarities in
         # range
-        TOOLS = dict(
-            tools='box_select, box_zoom, reset, pan, save, wheel_zoom, lasso_select')
+        TOOLS = dict(tools='box_select, box_zoom, reset, pan, save,\
+                            wheel_zoom, lasso_select')
         ax1 = figure(sizing_mode='scale_both', **TOOLS)
         ax2 = figure(sizing_mode='scale_both', x_range=ax1.x_range, **TOOLS)
+
+        stats_text = stats_display(tt, gain_type, doplot, corr, field)
+
+        # list for collecting plot states
+        ax1_plots = []
+        ax2_plots = []
 
         # forming Legend object items for data and errors
         legend_items_ax1 = []
@@ -985,22 +1409,17 @@ def main(**kwargs):
             # querying the table for the 2 columns
 
             subtab = tt.query(query=mytaql)
-            # Selecting values from the table for antenna ants
-            times = subtab.getcol('TIME')
-            # Referencing time wrt the initial time
-            times = times - times[0]
 
-            flagcol = subtab.getcol('FLAG')
-
-            paramerr = subtab.getcol('PARAMERR')
+            xdata, xlabel = get_xaxis_data(subtab, gain_type)
+            prepd_x = prep_xaxis_data(xdata, gain_type)
+            ydata, y1label, y2label = get_yaxis_data(subtab, gain_type,
+                                                     doplot)
 
             # for tooltips
-            spw_id = subtab.getcol('SPECTRAL_WINDOW_ID')
-            scan_no = subtab.getcol('SCAN_NUMBER')
-            sub_ants = subtab.getcol('ANTENNA1')
-            ttip_antnames = [antnames[x] for x in sub_ants]
+            spw_id, scan_no, ttip_antnames = get_tooltip_data(subtab,
+                                                              gain_type)
 
-            tab_tooltips = [("(x,y)", "($x,$y)"),
+            tab_tooltips = [("(x, y)", "($x, $y)"),
                             ("spw", "@spw"),
                             ("scan_id", "@scanid"),
                             ("antenna", "@antname")]
@@ -1010,51 +1429,16 @@ def main(**kwargs):
             hover2 = HoverTool(tooltips=tab_tooltips,
                                mode='mouse', point_policy='snap_to_data')
 
-            if gain_type is 'G':
-                cparam = subtab.getcol('CPARAM')
+            ax1.xaxis.axis_label = ax1_xlabel = xlabel
+            ax2.xaxis.axis_label = ax2_xlabel = xlabel
+            ax1.yaxis.axis_label = ax1_ylabel = y1label
+            ax2.yaxis.axis_label = ax2_ylabel = y2label
 
-                # creating a masked array to prevent invalid values from being
-                # computed. IF mask is true, then the element is masked, and thus
-                # won't be used in the calculation
-                # removing flagged data from cparams
-                masked_data = np.ma.masked_array(data=cparam, mask=flagcol)
-                masked_data_err = np.ma.masked_array(data=paramerr,
-                                                     mask=flagcol)
-
-                y1, y1_err, y2, y2_err = data_prep_G(masked_data,
-                                                     masked_data_err,
-                                                     doplot, corr)
-                # setting up glyph data source
-                source = ColumnDataSource(
-                    data=dict(x=times, y1=y1, y2=y2, spw=spw_id, scanid=scan_no, antname=ttip_antnames))
-                ax1.xaxis.axis_label = ax1_xlabel = 'Time [s]'
-                ax2.xaxis.axis_label = ax2_xlabel = 'Time [s]'
-
-            elif gain_type is 'B':
-                cparam = subtab.getcol('CPARAM')
-                nchan = cparam.shape[1]
-                chans = np.arange(0, nchan, dtype='int')
-
+            if gain_type is 'B':
+                nchan = get_frequencies(subtab).size
+                chans = np.arange(nchan)
                 # for tooltips
-                spw_id = spw_id.reshape(spw_id.size, 1)
-                spw_id = spw_id.repeat(nchan, axis=1)
-                scan_no = scan_no.reshape(scan_no.size, 1)
-                scan_no = scan_no.repeat(nchan, axis=1)
-                sub_ants = sub_ants.reshape(sub_ants.size, 1)
-                sub_ants = sub_ants.repeat(nchan, axis=1)
                 ttip_antnames = [antlabel] * nchan
-
-                masked_data = np.ma.masked_array(data=cparam, mask=flagcol)
-                masked_data_err = np.ma.masked_array(data=paramerr,
-                                                     mask=flagcol)
-
-                y1, y1_err, y2, y2_err = data_prep_B(masked_data,
-                                                     masked_data_err,
-                                                     doplot, corr)
-                source = ColumnDataSource(
-                    data=dict(x=chans, y1=y1, y2=y2, spw=spw_id, scanid=scan_no, antname=ttip_antnames))
-                ax1.xaxis.axis_label = ax1_xlabel = 'Channel'
-                ax2.xaxis.axis_label = ax2_xlabel = 'Channel'
 
                 if ant == plotants[-1]:
                     linax1 = add_axis(ax1, (frequencies[0], frequencies[-1]),
@@ -1064,54 +1448,32 @@ def main(**kwargs):
                     ax1.add_layout(linax1, 'above')
                     ax2.add_layout(linax2, 'above')
 
-            elif gain_type is 'K':
-                if doplot == 'ap':
-                    antenna = subtab.getcol('ANTENNA1')
-                    fparam = subtab.getcol('FPARAM')
-                    masked_data = np.ma.masked_array(data=fparam, mask=flagcol)
-                    masked_data_err = np.ma.masked_array(data=paramerr,
-                                                         mask=flagcol)
+            if gain_type is 'K':
+                if doplot == 'ri':
+                    logger.error('Exiting: No complex values to plot')
+                    # break #[for when there'r multiple tables to be plotted]
+                    sys.exit(-1)
 
-                    y1, y1_err, y2, y2_err = data_prep_K(
-                        masked_data, masked_data_err, corr)
-                    source = ColumnDataSource(
-                        data=dict(x=antenna, y1=y1, y2=y2, spw=spw_id, scanid=scan_no, antname=ttip_antnames))
-                    ax1.xaxis.axis_label = ax1_xlabel = 'Antenna'
-                    ax2.xaxis.axis_label = ax2_xlabel = 'Antenna'
-                else:
-                    logging.info('ragavi exited: No complex values to plot')
-                    sys.exit()
+            y1, y1_err, y2, y2_err = prep_yaxis_data(subtab, ydata,
+                                                     gain_type,
+                                                     ptype=doplot,
+                                                     corr=corr,
+                                                     flag=True)
 
-            elif gain_type is 'F':
-                cparam = subtab.getcol('CPARAM')
-                masked_data = np.ma.masked_array(data=cparam, mask=flagcol)
-                masked_data_err = np.ma.masked_array(data=paramerr,
-                                                     mask=flagcol)
-
-                y1, y1_err, y2, y2_err = data_prep_F(
-                    masked_data, masked_data_err, doplot, corr)
-                # setting up glyph data source
-                source = ColumnDataSource(
-                    data=dict(x=times, y1=y1, y2=y2, spw=spw_id, scanid=scan_no, antname=ttip_antnames))
-                ax1.xaxis.axis_label = ax1_xlabel = 'Time [s]'
-                ax2.xaxis.axis_label = ax2_xlabel = 'Time [s]'
+            source = ColumnDataSource(data=dict(x=prepd_x, y1=y1, y2=y2,
+                                                spw=spw_id, scanid=scan_no,
+                                                antname=ttip_antnames))
 
             p1, p1_err, p2, p2_err = make_plots(
                 source=source, color=y1col, ax1=ax1, ax2=ax2, y1_err=y1_err)
 
-            if doplot == 'ap':
-                ax1.yaxis.axis_label = ax1_ylabel = 'Amplitude'
-                ax2.yaxis.axis_label = ax2_ylabel = 'Phase [Deg]'
-                if gain_type is 'K':
-                    ax1.yaxis.axis_label = ax1_ylabel = 'Amplitude'
-                    ax2.yaxis.axis_label = ax2_ylabel = 'Amplitude'
-            elif doplot == 'ri':
-                ax1.yaxis.axis_label = ax1_ylabel = 'Real'
-                ax2.yaxis.axis_label = ax2_ylabel = 'Imaginary'
-
             # hide all the other plots until legend is clicked
             if ant > 0:
                 p1.visible = p2.visible = False
+
+            # collecting plot states for each iterations
+            ax1_plots.append(p1)
+            ax2_plots.append(p2)
 
             # forming legend object items
             legend_items_ax1.append((legend, [p1]))
@@ -1122,10 +1484,10 @@ def main(**kwargs):
 
             subtab.close()
 
-            if np.min(times) < xmin:
-                xmin = np.min(times)
-            if np.max(times) > xmax:
-                xmax = np.max(times)
+            if np.min(prepd_x) < xmin:
+                xmin = np.min(prepd_x)
+            if np.max(prepd_x) > xmax:
+                xmax = np.max(prepd_x)
             if np.min(y1) < yumin:
                 yumin = np.min(y1)
             if np.max(y1) > yumax:
@@ -1168,7 +1530,6 @@ def main(**kwargs):
         ax2.y_range = Range1d(ylmin, ylmax)
 
         tt.close()
-        spw_table.close()
 
         # configuring titles for the plots
         ax1_title = Title(text=ax1_ylabel + ' vs ' + ax1_xlabel,
@@ -1208,6 +1569,10 @@ def main(**kwargs):
         ax2.add_layout(ax2_title, 'above')
         ax1.add_layout(ax1_title, 'above')
 
+        # creating size slider for the plots
+        size_slider = Slider(end=10, start=1, step=0.5,
+                             value=4, title='Scatter point size')
+
         # creating and configuring Antenna selection buttons
         ant_select = Toggle(label='Select All Antennas',
                             button_type='success', width=200)
@@ -1221,7 +1586,7 @@ def main(**kwargs):
         batch_select = CheckboxGroup(labels=ant_labs, active=[])
 
         # Dropdown to hide and show legends
-        legend_toggle = Select(title="Showing Legends: ", value="alo",
+        legend_toggle = Select(title="Showing Legends: ", value="non",
                                options=[("all", "All"), ("alo", "Antennas"),
                                         ("elo", "Errors"), ("non", "None")])
 
@@ -1251,8 +1616,14 @@ def main(**kwargs):
                 loax2_err=legend_objs_ax2_err.values()),
             code=legend_toggle_callback())
 
-        plot_widgets = widgetbox(
-            [ant_select, batch_select, toggle_err, legend_toggle])
+        size_slider.callback = CustomJS(args={'slide': size_slider,
+                                              'p1': ax1_plots,
+                                              'p2': ax2_plots},
+                                        code=size_slider_callback())
+
+        plot_widgets = widgetbox([ant_select, batch_select,
+                                  toggle_err, legend_toggle,
+                                  stats_text, size_slider])
 
         if gain_type is not 'K':
             layout = gridplot([[plot_widgets, ax1, ax2]],
@@ -1266,6 +1637,7 @@ def main(**kwargs):
     if image_name:
         save_svg_image(image_name, ax1, ax2,
                        legend_items_ax1, legend_items_ax2)
+
     if not NB_RENDER:
         if html_name:
             save_html(html_name, final_layout)
@@ -1279,11 +1651,11 @@ def main(**kwargs):
             if len(mytabs) > 1:
                 mytab = datetime.now().strftime('%Y%m%d_%H%M%S')
             html_name = "{}_corr_{}_{}_field_{}".format(mytab, corr,
-                                                        doplot, fields)
-
+                                                        doplot,
+                                                        ''.join(field_ids))
             save_html(html_name, final_layout)
 
-        logging.info("Rendered: {}.html".format(html_name))
+        logger.info("Rendered: {}.html".format(html_name))
 
     else:
         output_notebook()
@@ -1293,7 +1665,7 @@ def main(**kwargs):
 
 def plot_table(mytabs, gain_types, fields, **kwargs):
     """
-    Function for plotting tables 
+    Function for plotting tables
 
     Inputs
     --------
@@ -1310,21 +1682,21 @@ def plot_table(mytabs, gain_types, fields, **kwargs):
 
         doplot      : Plot complex values as amp and phase (ap) or real and
                       imag (ri) (default = ap)',default='ap')
-        plotants    : Plot only this antenna, or comma-separated string of 
+        plotants    : Plot only this antenna, or comma-separated string of
                       antennas',default=[-1])
-        corr        : Correlation index to plot (usually just 0 or 1, 
+        corr        : Correlation index to plot (usually just 0 or 1,
                       default = 0)',default=0)
         t0          : Minimum time to plot (default = full range)',default=-1)
         t1          : Maximum time to plot (default = full range)',default=-1)
-        yu0         : Minimum y-value to plot for upper panel 
+        yu0         : Minimum y-value to plot for upper panel
                       (default = full   range)',default=-1)
-        yu1         : Maximum y-value to plot for upper panel 
+        yu1         : Maximum y-value to plot for upper panel
                       (default = full range)',default=-1)
-        yl0         : Minimum y-value to plot for lower panel 
+        yl0         : Minimum y-value to plot for lower panel
                       (default = full range)',default=-1)
-        yl1         : Maximum y-value to plot for lower panel 
+        yl1         : Maximum y-value to plot for lower panel
                       (default = full range)',default=-1)
-        mycmap      : Matplotlib colour map to use for antennas 
+        mycmap      : Matplotlib colour map to use for antennas
                       (default = coolwarm)',default='coolwarm')
         image_name  : Output image name (default = something sensible)'
 
@@ -1336,7 +1708,7 @@ def plot_table(mytabs, gain_types, fields, **kwargs):
     """
     if mytabs is None:
         print("Please specify a gain table to plot.")
-        logging.info('ragavi exited: No gain table specfied.')
+        logger.error('Exiting: No gain table specfied.')
         sys.exit(-1)
     else:
         # getting the name of the gain table specified